# internal/function: add cache for history_idx